requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pywebview==5.3.2
jinja2==3.1.5
python-multipart==0.0.20
aiofiles==24.1.0
keyboard==0.13.5
